+++ SpaceCapsule/SpaceCapsule/Model/JWT/Header.swift
@@ -1,0 +1,13 @@
+//
+//  Header.swift
+//  SpaceCapsule
+//
+//  Created by young june Park on 2022/12/12.
+//
+
+import Foundation
+
+struct Header: Encodable {
+    let alg = "ES256"
+    let kid = "4979Z7MUXJ"
+}
